Adds DoubleMapper
DoubleMapper is able to match a pair of consecutive abbreviations either
to the left or right side. If no pair is provided, it does not match any
entry.

AbstractMapper is adapted to provide a DRY method for mapping. No
changes are expected for SingleMapper. For DoubleMapper, the context
should be checked for valid matches.